Make Carmanufacturingcontroller Great Again!
(Make CarManufacturingController observe workstations instead of assemAssist)
