+++ src/swop/Exceptions/NotAllTasksCompleteException.java
@@ -9,13 +9,4 @@ public class NotAllTasksCompleteException extends Exception {
         super(errorMessage);
         this.workstations = w;
     }
-
-    public String getWorkstations() {
-    	String all = "";
-    	for(String w : this.workstations) {
-    		all += w;
-    		all += ", ";
-    	}
-        return all.substring(0, all.length() - 2);
-    }
 }
